+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
-### unreleased
-- Remove unused Minterms method of Kmap
+### [v1.0.0](https://github.com/noah-friedman/kmap/tree/v1.0.0)
+- Implement don't care conditions
+- Remove unused Minterms method of [Kmap](https://github.com/noah-friedman/kmap/blob/v1.0.0/kmap.go)
 
 ### [v0.1.1](https://github.com/noah-friedman/kmap/tree/v0.1.1)
 - [Error messages](https://github.com/noah-friedman/kmap/blob/v0.1.1/kmap.go) explain what went wrong better